+++ app/src/main/java/com/piwniczna/mojakancelaria/activities/other/settings/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.piwniczna.mojakancelaria.activities.other
+package com.piwniczna.mojakancelaria.activities.other.settings
 
 
 import android.os.AsyncTask
