+++ app/src/main/res/layout/layout_dialog_get_backup_name.xml
@@ -25,7 +25,7 @@
 
         android:text="@string/ok"
         android:layout_weight="0"
-        style="@style/small_button_style"
+        style="@style/small_dialog_button_style"
         android:layout_gravity="center"
         android:backgroundTint="@color/dark_blue"
         android:id="@+id/confirm_backup"
