+++ app/src/main/res/layout/layout_dialog_new_avizo.xml
@@ -57,7 +57,7 @@
 
         android:text="@string/ok"
         android:layout_weight="0"
-        style="@style/small_button_style"
+        style="@style/small_dialog_button_style"
         android:layout_gravity="center"
         android:backgroundTint="@color/dark_blue"
         android:id="@+id/confirm_new_letter"
